jstl 의존성 추가, Clob -> Lob 변경, 로그 방식 변경, 게시판(공지사항/예매오픈안내) 기본 세팅

+++ src/main/java/com/heeji/picket/domain/Notification.java
@@ -17,11 +17,11 @@ public class Notification {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Comment("알림고유번호")
-    private int noticeId;
+    private Long noticeId;
 
     @Column(nullable = false)
     @Comment("공연정보고유번호")
-    private int showId;
+    private Long showId;
 
     @Column(nullable = false)
     @Comment("알림유형")
