+++ src/main/java/com/heeji/picket/domain/Seat.java
@@ -17,19 +17,19 @@ public class Seat {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Comment("좌석고유번호")
-    private int seatId;
+    private Long seatId;
 
     @Column(nullable = false, unique = true)
     @Comment("좌석등급고유번호")
-    private int seatGradeId;
+    private Long seatGradeId;
 
     @Column(nullable = false, unique = true)
     @Comment("공연날짜고유번호")
-    private int showDateId;
+    private Long showDateId;
 
     @Column(nullable = false)
     @Comment("공연정보고유번호")
-    private int showId;
+    private Long showId;
 
     @Column(nullable = false, unique = true)
     @Comment("열")
