+++ src/main/java/com/heeji/picket/domain/SeatGrade.java
@@ -4,11 +4,6 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.hibernate.annotations.Comment;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.sql.Clob;
-import java.sql.Timestamp;
 
 @Entity
 @AllArgsConstructor
@@ -18,15 +13,15 @@ public class SeatGrade {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Comment("좌석등급고유번호")
-    private int seatGradeId;
+    private Long seatGradeId;
 
     @Column(nullable = false, unique = true)
     @Comment("공연날짜고유번호")
-    private int showDateId;
+    private Long showDateId;
 
     @Column(nullable = false)
     @Comment("공연정보고유번호")
-    private int showId;
+    private Long showId;
 
     @Column(nullable = false)
     @Comment("좌석개수")
