+++ src/main/java/com/heeji/picket/domain/ShowDate.java
@@ -5,9 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.hibernate.annotations.Comment;
 import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
 
-import java.sql.Clob;
 import java.sql.Timestamp;
 
 @Entity
@@ -18,11 +16,11 @@ public class ShowDate {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Comment("공연날짜고유번호")
-    private int showDateId;
+    private Long showDateId;
 
     @Column(nullable = false, unique = true)
     @Comment("공연정보고유번호")
-    private int showId;
+    private Long showId;
 
     @CreationTimestamp
     @Column(nullable = false, unique = true)
